pyproject.toml
[tool.poetry]
name = "lightweight-ai-agent"
version = "0.1.0"
description = "A lightweight AI agent system for managing calendar, reading emails, and searching events."
authors = ["Your Name <youremail@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.11.0"
python-dotenv = "^1.0.1"
fastapi = "^0.115.12"
uvicorn = "^0.34.0"
python-multipart = "^0.0.20"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"